Move TeaserOrgTeamsRow into its own component
