+++ packages/client/modules/userDashboard/components/OrgTeams/OrgTeams.tsx
@@ -1,8 +1,6 @@
-import {Lock} from '@mui/icons-material'
 import graphql from 'babel-plugin-relay/macro'
 import React from 'react'
 import {useFragment} from 'react-relay'
-import {useHistory} from 'react-router'
 import {OrgTeams_organization$key} from '../../../../__generated__/OrgTeams_organization.graphql'
 import AddTeamDialogRoot from '../../../../components/AddTeamDialogRoot'
 import {Button} from '../../../../ui/Button/Button'
@@ -35,8 +33,6 @@ const OrgTeams = (props: Props) => {
     organizationRef
   )
 
-  const history = useHistory()
-
   const {
     open: openAddTeamDialog,
     close: closeAddTeamDialog,
@@ -46,10 +42,6 @@ const OrgTeams = (props: Props) => {
   const {allTeams, tier, viewerTeams, allTeamsCount} = organization
   const showAllTeams = allTeams.length === allTeamsCount
   const viewerTeamCount = viewerTeams.length
-
-  const handleSeePlansClick = () => {
-    history.push(`/me/organizations/${organization.id}/billing`)
-  }
 
   return (
     <div className='max-w-4xl pb-4'>
@@ -84,26 +76,10 @@ const OrgTeams = (props: Props) => {
         </div>
 
         {tier !== 'enterprise' && allTeamsCount > viewerTeamCount && !showAllTeams && (
-          <>
-            <TeaserOrgTeamsRow hiddenTeamCount={allTeamsCount - viewerTeamCount} />
-            <div className='flex items-center justify-between bg-white px-4 pb-4'>
-              <div className='flex items-center'>
-                <Lock className='h-10 w-10 select-none rounded-full p-1.5 text-grape-500' />
-                <p className='ml-3 text-sm text-slate-700'>
-                  Parabol Enterprise includes our Org Admin role, which allows you to see{' '}
-                  <strong>all</strong> teams in your organization
-                </p>
-              </div>
-              <Button
-                variant='destructive'
-                shape='pill'
-                className='w-32 py-2 text-base'
-                onClick={handleSeePlansClick}
-              >
-                See plans
-              </Button>
-            </div>
-          </>
+          <TeaserOrgTeamsRow
+            hiddenTeamCount={allTeamsCount - viewerTeamCount}
+            orgId={organization.id}
+          />
         )}
       </div>
